move text map getters inside trace context
