+++ lib/opentelemetry/trace/propagation/trace_context/text_map_getter.lua
